server/pyproject.toml
[project]
name = "buddy"
version = "0.1.0"
description = "Voice companion — local-first Pipecat voice AI agent"
requires-python = ">=3.10,<3.14"
dependencies = [
    "pipecat-ai[anthropic,kokoro,runner,silero,webrtc]",
    "python-dotenv",
    "loguru",
    "aiohttp",
]

[tool.setuptools]
py-modules = ["bot", "config", "stt_whisper", "tts_piper"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
